refactor: rename product-card to product-card-base
